PDFBOX-4892: Sonar fix
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1890725 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/rendering/GroupGraphics.java
@@ -658,7 +658,7 @@ class GroupGraphics extends Graphics2D
                     }
 
                     // Alpha factor alpha0 / alphagn - alpha0 is in range 0.0-1.0.
-                    float alphaFactor = alpha0 / (float)alphagn - alpha0 / 255.0f;
+                    float alphaFactor = alpha0 / alphagn - alpha0 / 255.0f;
                     int groupRGB = groupData[index]; // color of group pixel
 
                     // Compute backdrop removal for RGB components.
